+++ packages/medical-domain/test/repositories/medical/medical-data.service.test.ts
@@ -117,6 +117,7 @@ const wizardDeduplicateMock = jest.fn(
     return data
   }
 )
+
 const physicalActivityDeduplicateMock = jest.fn(
   (data: PhysicalActivity[], _opts: MedicalDataOptions) => {
     return data